move pages to their respective folder
