+++ frontend/src/App.tsx
@@ -1,6 +1,6 @@
-import AuthLayout from "./_auth/AuthLayout";
-import SigninForm from "./_auth/forms/SigninForm";
-import SignupForm from "./_auth/forms/SignupForm";
+import AuthLayout from "./pages/AuthLayout";
+import SigninForm from "./pages/SigninForm";
+import SignupForm from "./pages/SignupForm";
 import { AuthProvider } from "./contexts/AuthContext";
 import Landing from "./pages/Landing";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
