+++ frontend/src/pages/AuthLayout.tsx
