+++ frontend/src/pages/SigninForm.tsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import Input from "../../components/Input";
+import Input from "../components/Input";
 
 const SigninForm = () => {
   return (
